refactor: move config and event bus modules
