+++ src/ai_karen_engine/config/__init__.py
@@ -1,6 +1,6 @@
 """Configuration helpers and model registry utilities."""
 
-from config import model_registry
+from ai_karen_engine.config import model_registry
 
 __all__ = ["model_registry"]
