+++ src/ai_karen_engine/config/model_registry.py
