+++ src/ai_karen_engine/config/plugin_schema.json
