+++ src/ai_karen_engine/config/settings.json
